Only display include file notice if the file exists
Fixes #34

+++ premium/classes/admin/pages/class-page-redirect.php
@@ -28,17 +28,19 @@ class WPSEO_Page_Redirect {
 		<div class="tabwrapper>">
 			<div id="redirects" class="wpseotab">
 				<?php
-				echo '<div style="margin: 5px 0; padding: 3px 10px; background-color: #ffffe0; border: 1px solid #E6DB55; border-radius: 3px">';
-				if ( wpseo_is_nginx() ) {
-					echo '<p>' . __( "As you're on NGINX, you should add the following include to the website Nginx config file:", 'wordpress-seo' ) . '</p>';
-					echo '<pre>include ' . WPSEO_Redirect_File_Manager::get_file_path() . ';</pre>';
-				} else {
-					if ( wpseo_is_apache() ) {
-						echo '<p>' . __( "As you're on Apache, you should add the following include to the website httpd config file:", 'wordpress-seo' ) . '</p>';
-						echo '<pre>Include ' . WPSEO_Redirect_File_Manager::get_file_path() . '</pre>';
+				if ( file_exists( WPSEO_Redirect_File_Manager::get_file_path() ) ) {
+					echo '<div style="margin: 5px 0; padding: 3px 10px; background-color: #ffffe0; border: 1px solid #E6DB55; border-radius: 3px">';
+					if ( wpseo_is_nginx() ) {
+						echo '<p>' . __( "As you're on NGINX, you should add the following include to the website Nginx config file:", 'wordpress-seo' ) . '</p>';
+						echo '<pre>include ' . WPSEO_Redirect_File_Manager::get_file_path() . ';</pre>';
+					} else {
+						if ( wpseo_is_apache() ) {
+							echo '<p>' . __( "As you're on Apache, you should add the following include to the website httpd config file:", 'wordpress-seo' ) . '</p>';
+							echo '<pre>Include ' . WPSEO_Redirect_File_Manager::get_file_path() . '</pre>';
+						}
 					}
+					echo '</div>';
 				}
-				echo '</div>';
 
 				// Add new redirect HTML
 				echo "<form id='wpseo-new-redirects-form' method='post'>\n";
